requirements.txt
absl-py==1.4.0
Flask==2.3.3
Flask-Cors==4.0.0
mediapipe==0.10.9
numpy==1.24.3
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
pandas==2.0.3
tensorflow==2.13.1
Werkzeug==2.3.7
gunicorn==21.2.0
